% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wafflerize.R
\name{get_resolutions}
\alias{get_resolutions}
\title{Get resolutions of regular point grid}
\usage{
get_resolutions(x)
}
\arguments{
\item{x}{dataframe with x/y coordinates, preferably from regular point grid}
}
\description{
Get resolutions of regular point grid
}
