% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wafflerize.R
\name{wafflerize}
\alias{wafflerize}
\title{Create multipolygons waffle from regular point grid}
\usage{
wafflerize(x, fact = 100, res_x, res_y)
}
\arguments{
\item{x}{dataframe with x/y coordinates, preferably from regular point grid}

\item{fact}{multiplicative factor to magnify the coordinates so that it is not too small for mapping}

\item{res_x}{x resolution of the regular grid}

\item{res_y}{y resolution of the regular grid}
}
\value{
polygon
}
\description{
Create multipolygons waffle from regular point grid
}
